Include resource path in title
Previously the stylesheet would use either the resource doc title or the
resource path (with parameters appended).  Although this gives the WADL
author more flexibility for exactly what is in the title, it leads to
annoying duplication if the author prefers to have the path in the
title.  Therefore, add the path to the title automatically.

Note:  This is a style choice and may not be generally preferable.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>